Only symlink if you need to

+++ link.sh
@@ -47,7 +47,6 @@ function link_files() {
     done
 }
 
-
 function setup_oh_my_zsh {
     install_oh_my_zsh
     setup_oh_my_zsh_theme
@@ -62,8 +61,11 @@ function install_oh_my_zsh {
 
 function setup_oh_my_zsh_theme {
     CUSTOM_THEME_DIR=$HOME/.oh-my-zsh/custom/themes
-    mkdir -p $CUSTOM_THEME_DIR
-    ln -vs $DOTFILES_DIRECTORY/aseles.zsh-theme $CUSTOM_THEME_DIR/aseles.zsh-theme
+
+    if [ ! -L "$CUSTOM_THEME_DIR/aseles.zsh-theme" ] ; then
+        mkdir -p $CUSTOM_THEME_DIR
+        ln -vs $DOTFILES_DIRECTORY/aseles.zsh-theme $CUSTOM_THEME_DIR/aseles.zsh-theme
+    fi
 }
 
 main
